refactor: drop browser support
