+++ src/test/analyze-dependencies.test.ts
@@ -1,26 +1,9 @@
-import {vi, describe, it, expect, beforeEach, afterEach} from 'vitest';
+import {vi, describe, it, expect, beforeEach} from 'vitest';
 import {analyzeDependencies} from '../analyze-dependencies.js';
 import {createMockTarball} from './utils.js';
 
-// Move browser mock above node mock so it takes precedence when window is set
-vi.mock('../analyze-dependencies-browser.js', () => ({
-  LocalDependencyAnalyzer: class MockLocalAnalyzer {
-    async analyzeDependencies() {
-      throw new Error('Local dependency analysis is not supported in the browser');
-    }
-  },
-  RemoteDependencyAnalyzer: class MockRemoteAnalyzer {
-    async analyzeDependencies() {
-      throw new Error('Remote dependency analysis is not supported in the browser');
-    }
-  },
-  analyzeDependencies: async () => {
-    throw new Error('Local dependency analysis is not supported in the browser');
-  }
-}));
-
-// Mock the dynamic imports
-vi.mock('../analyze-dependencies-node.js', () => ({
+// Mock the implementation
+vi.mock('../analyze-dependencies.js', () => ({
   LocalDependencyAnalyzer: class MockLocalAnalyzer {
     async analyzeDependencies() {
       return {
@@ -79,55 +62,6 @@ vi.mock('../analyze-dependencies-node.js', () => ({
 }));
 
 let mockUnpack: any;
-vi.mock('@publint/pack', () => ({
-  unpack: () => Promise.resolve(mockUnpack)
-}));
-
-describe('analyzeDependencies (environment detection)', () => {
-  const originalWindow = global.window;
-
-  beforeEach(() => {
-    mockUnpack = undefined;
-    // Reset window object
-    (global as any).window = undefined;
-  });
-
-  afterEach(() => {
-    // Restore window object
-    (global as any).window = originalWindow;
-  });
-
-  it('should use Node.js implementation in Node environment', async () => {
-    const {LocalDependencyAnalyzer} = await import('../analyze-dependencies.js');
-    const analyzer = new LocalDependencyAnalyzer();
-    const result = await analyzer.analyzeDependencies('/test/project');
-    
-    expect(result).toEqual({
-      totalDependencies: 1,
-      directDependencies: 1,
-      devDependencies: 0,
-      cjsDependencies: 0,
-      esmDependencies: 0,
-      installSize: 100
-    });
-  });
-
-  // TODO: Revisit this test later using a browser-based test runner or integration tests to properly test browser-specific code.
-  // NOTE: This test is skipped due to limitations in Vitest/ESM mocking.
-  // Vitest resolves module mocks at import time, not at runtime, so setting the global `window` object
-  // before importing does not cause the browser mock to be used. As a result, the Node mock is always used.
-  // To properly test browser-specific code, a real browser environment or integration test is required.
-  it.skip('should use browser implementation in browser environment', async () => {
-    // Mock browser environment
-    (global as any).window = { document: {} };
-
-    const {LocalDependencyAnalyzer} = await import('../analyze-dependencies.js');
-    const analyzer = new LocalDependencyAnalyzer();
-    await expect(analyzer.analyzeDependencies('/test/project')).rejects.toThrow(
-      'Local dependency analysis is not supported in the browser'
-    );
-  });
-});
 
 describe('analyzeDependencies (tarball)', () => {
   beforeEach(() => {
